homework2-localfiles/requirements.txt
anyio==4.3.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
finnhub-python==2.4.19
Flask==3.0.2
Flask-Cors==4.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
MarkupSafe==2.1.5
packaging==23.2
polygon==1.2.0
polygon-api-client==1.13.4
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
urllib3==1.26.18
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
